fix: [BOJ] update 27433.cpp

+++ PS/BOJ/20001-30000/27001-28000/27433.cpp
@@ -2,7 +2,7 @@
 
 using namespace std;
 
-int calc(long long x) {
+long long calc(long long x) {
     if(x == 0) return 1;
     else if(x > 1) return x * calc(x - 1);
     else return x;
